requirements.txt
blinker==1.9.0
captcha==0.6.0
click==8.1.8
Flask==3.1.0
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.1
pillow==11.0.0
PyJWT==2.10.1
Werkzeug==3.1.3
WTForms==3.2.1
